+++ buildSrc/src/main/kotlin/SamplesDependencies.kt
@@ -3,11 +3,11 @@ object SamplesVersions {
     val androidPlugin = PluginVersions.androidPlugin
     val kotlinNativePlugin = PluginVersions.kotlinNativePlugin
     val gradleWrapper = PluginVersions.gradleWrapper
-    val ktlintCore = "0.22.0"
+    val ktlintCore = "0.23.1"
+    val junit = PluginVersions.junit
 
     // Android sample dependencies
     val androidSupport = "27.0.2"
-    val junit = "4.12"
     val espressoRunner = "1.0.1"
     val espresso = "3.0.1"
 }