+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased [4.1.0-SNAPSHOT]
+### Added
+  - ?
+### Changed
+  - ?
+### Removed
+  - ?
+### Fixed
+  - ?
+
 ## [4.0.0] - 2018-5-15
 ### Added
  - Add support for providing custom rulesets (#71)