+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.2.0-SNAPSHOT] - Unreleased
+### Added
+  - ?
+### Changed
+  - ?
+### Removed
+  - ?
+### Fixed
+  - ?
+
 ## [4.1.0] - 2018-6-13
 ### Added
   - apply to IDEA task (in the project) (#91)